Add selector configuration

+++ massa-pos-exports/src/error.rs
@@ -27,7 +27,7 @@ pub enum PosError {
     CannotComputeSeed,
     /** Invalid number of initial rolls, {0} found and we expect a number >=
     loop to the loopback_cycle number {1} */
-    InvalidInitialRolls(usize, usize),
+    InvalidInitialRolls(u64, usize),
     /// serde error
     SerdeError(#[from] serde_json::Error),
     /// Io error
